+++ internal/cnitap/types.go
@@ -12,8 +12,8 @@ package cnitap
 import (
 	"github.com/containernetworking/cni/pkg/types"
 
-	"go.datum.net/galactic/internal/cni/hostconf"
 	"go.datum.net/galactic/internal/cniipam"
+	"go.datum.net/galactic/internal/hostconf"
 )
 
 // PluginConf is the CNI plugin configuration passed via stdin on each
