+++ internal/crdnames/crdnames_test.go
