+++ internal/nadpatch/nadpatch_test.go
